Financial Calculator
It is my learning project. I learnt using pandas using this. This project was requested by my cousin brother and I tried making it. In this project I asked the user to enter:
expense name:
total cost of that expenditure:
and save it to a excel file. later you can view the excel file, change it, delete it etc. it was just for learning of pandas